Redshift: only drop dependent views if they exist (#614)

+++ parsons/databases/redshift/redshift.py
@@ -920,9 +920,10 @@ class Redshift(RedshiftCreateTable, RedshiftCopyTable, RedshiftTableUtilities, R
             connection.set_session(autocommit=True)
             tbl = self.query_with_connection(sql_depend, connection)
             dropped_views = [row['table_name'] for row in tbl]
-            sql_drop = "\n".join([f"drop view {view};" for view in dropped_views])
-            tbl = self.query_with_connection(sql_drop, connection)
-            logger.info(f"Dropped the following views: {dropped_views}")
+            if dropped_views:
+                sql_drop = "\n".join([f"drop view {view};" for view in dropped_views])
+                tbl = self.query_with_connection(sql_drop, connection)
+                logger.info(f"Dropped the following views: {dropped_views}")
 
         return tbl
